Address the-dijkstra comments
